+++ src/REvoDesign/phylogenetics/EvoMutator.py
@@ -8,14 +8,12 @@ from functools import partial
 from typing import List, Literal, Optional, Tuple, Union
 
 import Bio.PDB.PDBParser as PDBParser
+import matplotlib
+import pandas as pd
 from immutabledict import immutabledict
 from joblib import Parallel, delayed
-import pandas as pd
 from pymol import CmdException, cmd
 from RosettaPy.common.mutation import Mutation, RosettaPyProteinSequence
-import matplotlib
-
-matplotlib.use("Agg")
 
 from REvoDesign import ConfigBus, issues
 from REvoDesign.basic import IterableLoop
@@ -39,6 +37,9 @@ from REvoDesign.tools.pymol_utils import (any_posision_has_been_selected,
                                           make_temperal_input_pdb)
 from REvoDesign.tools.utils import (cmap_reverser, get_color, rescale_number,
                                     run_worker_thread_with_progress, timing)
+
+matplotlib.use("Agg")
+
 
 logging = root_logger.getChild(__name__)
 
@@ -997,15 +998,15 @@ class GREMLIN_Analyser:
     def load_co_evolving_pairs(
         self,
         walk_to_next=True,
-    ): 
+    ):
         def mutate_with_gridbuttons(
-        col,
-        row,
-        matrix:pd.DataFrame,
-        min_score,
-        max_score,
-        pair: CoevolvedPair,
-        ignore_wt=False,
+            col,
+            row,
+            matrix: pd.DataFrame,
+            min_score,
+            max_score,
+            pair: CoevolvedPair,
+            ignore_wt=False,
         ):
             self.refresh_magician()
             self.picked_gremlin_group_id = "_vs_".join(
@@ -1074,8 +1075,8 @@ class GREMLIN_Analyser:
 
             # call scorer to evaluate wt and mutant
             if not self.magician.magician:
-                wt_score = matrix.loc[wt_i,wt_j]
-                mut_score = matrix.loc[mut_i,mut_j]
+                wt_score = matrix.loc[wt_i, wt_j]
+                mut_score = matrix.loc[mut_i, mut_j]
             else:
                 if self.magician.magician.no_need_to_score_wt:
                     wt_score = 0
@@ -1097,7 +1098,7 @@ class GREMLIN_Analyser:
             self.picked_gremlin_mutant = mutant_obj
 
             # if mutant obj exists, activate it.
-            if mutant_obj in self.explored_mutant_tree.all_mutant_objects:
+            if self.explored_mutant_tree.has(mutant_obj):
                 logging.info(
                     f"Picked mutant: {mutant_obj.short_mutant_id} ({mutant_obj.full_mutant_id}) already exists. Do nothing."
                 )
@@ -1159,7 +1160,7 @@ class GREMLIN_Analyser:
                 {self.picked_gremlin_group_id: {mutant: mutant_obj}}
             )
             self.to_broadcaster(mutant_tree)
-            
+
         with hold_trigger_button(
             buttons=self.bus.buttons(button_ids=("previous", "next"))
         ):
@@ -1209,12 +1210,12 @@ class GREMLIN_Analyser:
                 pair_i=pair.i,
                 pair_j=pair.j,
                 cmap=self.bus.get_value("ui.header_panel.cmap.default"))
-            
-            button_matrix.alphabet_col=list("ARNDCQEGHILKMFPSTWYV-")
-            button_matrix.alphabet_row=list("ARNDCQEGHILKMFPSTWYV-")
+
+            button_matrix.alphabet_col = list("ARNDCQEGHILKMFPSTWYV-")
+            button_matrix.alphabet_row = list("ARNDCQEGHILKMFPSTWYV-")
 
             button_matrix.init_ui()
-            button_matrix.active_func=partial(
+            button_matrix.active_func = partial(
                 mutate_with_gridbuttons,
                 matrix=button_matrix.df_matrix,
                 min_score=button_matrix.min_value,
@@ -1260,13 +1261,11 @@ class GREMLIN_Analyser:
             )
         else:
             logging.debug(
-                f" Rejecting  co-evolved mutant {picked_gremlin_mutant_id}"
+                f" Rejecting co-evolved mutant {picked_gremlin_mutant_id}"
             )
             cmd.disable(picked_gremlin_mutant_id)
-            if (
-                picked_gremlin_mutant_id
-                not in self.mutant_tree_coevolved.all_mutant_ids
-            ):
+            if (not self.mutant_tree_coevolved.has(picked_gremlin_mutant_id)
+                ):
                 logging.warning(
                     f"{picked_gremlin_mutant_id} has not been accepted yet. Skipped."
                 )
@@ -1312,10 +1311,8 @@ class GREMLIN_Analyser:
                 "Co-evolved pairs are not loaded. "
             )
 
-        if (
-            self.picked_gremlin_mutant
-            in self.explored_mutant_tree.all_mutant_objects
-        ):
+        if (self.explored_mutant_tree.has(self.picked_gremlin_mutant)
+            ):
             logging.warning(
                 f"Igore repetative picking: {self.picked_gremlin_mutant.short_mutant_id} ({self.picked_gremlin_mutant.full_mutant_id})"
             )
@@ -1366,7 +1363,6 @@ class GREMLIN_Analyser:
         self.magician = magician
 
         return
-
 
     def to_broadcaster(self, mutant_tree: MutantTree):
         if (